+++ dragonchain/webserver/lib/verifications.py
@@ -14,6 +14,7 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 # KIND, either express or implied. See the Apache License for the specific
 # language governing permissions and limitations under the Apache License.
+import re
 from typing import Dict, Union, List, Any, Optional, cast
 
 from dragonchain.broadcast_processor import broadcast_functions
@@ -24,6 +25,8 @@ from dragonchain import exceptions
 from dragonchain import logger
 
 _log = logger.get_logger()
+
+_uuid_regex = re.compile(r"[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}")
 
 
 def get_pending_verifications_v1(block_id: str) -> Dict[str, List[str]]:
@@ -61,7 +64,12 @@ def query_interchain_broadcasts_v1(block_id: str) -> List[Any]:
         timestamp = l5_block["header"]["timestamp"]
         dc_id = l5_block["header"]["dc_id"]
         l5_nodes = redisearch._get_redisearch_index_client(redisearch.Indexes.verification.value).redis.smembers(redisearch.L5_NODES)
-        results = [_query_l5_verification(l5_dc_id.decode("utf-8"), timestamp) for l5_dc_id in l5_nodes if l5_dc_id.decode("utf-8") != dc_id]
+
+        results = [
+            _query_l5_verification(l5_dc_id.decode("utf-8"), timestamp)
+            for l5_dc_id in l5_nodes
+            if l5_dc_id.decode("utf-8") != dc_id and not re.match(_uuid_regex, l5_dc_id.decode("utf-8"))
+        ]
     return ([l5_block] if l5_block else []) + [storage.get_json_from_object(f"BLOCK/{x}") for x in results if x is not None]