+++ dragonchain/broadcast_processor/broadcast_processor.py
@@ -223,7 +223,7 @@ async def send_notification_verification(
         )
         _log.debug(f"Notification <- {resp.status} {url}")
     except Exception:
-        _log.exception(f"Unable to send verification notification.")
+        _log.exception("Unable to send verification notification.")
 
     await broadcast_functions.remove_notification_verification_for_broadcast_async(redis_list_value)
 
@@ -245,6 +245,9 @@ async def process_blocks_for_broadcast(session: aiohttp.ClientSession) -> None: 
     for block_id, score in await broadcast_functions.get_blocks_to_process_for_broadcast_async():
         _log.info(f"[BROADCAST PROCESSOR] Checking block {block_id}")
         current_level = await broadcast_functions.get_current_block_level_async(block_id)
+        if current_level == -1:
+            _log.warning(f"Failed to lookup current level for block {block_id}.")
+            continue
         try:
             claim: Any = matchmaking.get_or_create_claim_check(block_id, _requirements)
         except exceptions.InsufficientFunds: