+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Update redis in helm chart to 6.0.5
 - **Bugs**
   - Fixed bug in broadcast processor causing failure when a block cannot be found in redis
+  - Fixed bug in querying interchain broadcasts when the old L5 chain ID is a UUID and not public key
 - **Development:**
   - Use helm 3.2.4 and yq 3.3.2 for dependency container and lint checking